refactor: create common.py
