+++ src/discord_avatar.py
@@ -6,10 +6,10 @@ import os
 import requests
 
 # local imports
-from discord_helpers import get_bot_avatar
+import common
 
 # avatar
-avatar = get_bot_avatar(gravatar=os.environ['GRAVATAR_EMAIL'])
+avatar = common.get_bot_avatar(gravatar=os.environ['GRAVATAR_EMAIL'])
 
 avatar_response = requests.get(url=avatar)
 avatar_img = BytesIO(avatar_response.content).read()
