+++ src/discord_helpers.py
@@ -2,7 +2,6 @@
 from typing import Union
 
 # lib imports
-from libgravatar import Gravatar
 import requests
 
 # convert month number to igdb human-readable month
@@ -20,29 +19,6 @@ month_dictionary = {
     11: 'Nov',
     12: 'Dec'
 }
-
-
-def get_bot_avatar(gravatar: str) -> str:
-    """
-    Get Gravatar image url.
-
-    Return the Gravatar image url of the given email address.
-
-    Parameters
-    ----------
-    gravatar : str
-        The Gravatar email address.
-
-    Returns
-    -------
-    str
-        Gravatar image url.
-    """
-
-    g = Gravatar(email=gravatar)
-    image_url = g.get_image()
-
-    return image_url
 
 
 def igdb_authorization(client_id: str, client_secret: str) -> dict:
